Added image support
Each node can have one image, selected from a dropdown of all images in the blend file, the images are Blender's auto-generated thumbnail file previews, and so are pretty low-res. If an image you want to use is not in your blend file already, there is an 'Open Image' button for you to add it.
These options can all be hidden locally (single node) in the Node Properties panel or globally (all nodes at once) in the Addon Preferences.
All hidden options are always available via the Node Properties panel.