Clean up
Removed unused imports, added the 'Socket' and 'Update' shortcuts to the node, and tidied up the layout in the properties panel to maintain consistency. Removed an unused variable or two.

Made a few attempts at generating String props on the fly to allow the user to fill in the socket names, but no joy, it may become a job for someone else. While it was easy enough to add or subtract sockets, adding and removing String properties dynamically at the same time is a bit beyond me, at the moment.