Fixed Missing Images bug
Using just an image preview was not enough to cause Blender to save the image data block upon restart.

I had thought this would be saved in the node tree, or the blend file, but it turns out, if the images aren't loaded into a texture node, then they're considered unused, and purged when Blender is restarted. I tried packing them automatically, manually, saving them, opening them in another 2d editing workspace (Texture Paint), but I could not get Blender to consider them "in use". 

That's about when I discovered fake users, and suddenly realized why everyone has been saying they're so important for so long.
I tried to set fake user to true at the same time as the picture is set in the node, but it didn't work, so I put it in the update method instead.

Really annoying bug that I only discovered after coming back to my test file the next day.