2K Images, t3dn_bip, streamlined workflow
Hi-Res image support has been added c/o t3dn_bip - The 3D Ninjas Blender Image Preview. 
Images will be rendered up to 2K (2048 x 2048), and can be selected via a preview enum.
The enum will be populated by a folder selected by the user, by default it will be the 'images' folder of the addon, but you can specify any image directory on your device and it will autopopulate the enum, and when the enum is updated it will place the hi-res image in the main image section of the node.
The folder can be selected in the Addon prefs, or in each individual node.
They all share the same folder, but can be set to different images within that folder.
Changing the folder on one node will allow you to add different images, and it will not affect the nodes you have already set up, however, if you attempt to change any of the old images, you will only have access to the current image folder.