Added Numberlist node
Another new node, a simple numbered list.
The alignment is off with item 10 (and above, when implemented). I could choose the standard 'EXPAND' alignment, which would align the 10th box, but it also sacrifices a huge amount of horizontal space in the node, whereas all the other alignment options will fill the horizontal space, but cause the 10th item to be slightly (some might say infuriatingly) offset.