BigText draw cleanup
Given the lack of context for node initialization, and the fact that you can't call operators during the node's draw function I abandoned trying to get a ref to the node tree and ended up cleaning up the draw conditions instead. (Did attempt to do it in the node's update function, but it ended up messy and didn't work any better in the end)
It still needs a cleanup in the unregister function that will remove active draw handlers on uninstall, but otherwise it seems to be working without issue.